Pig Latin Challenge
Given a piece of English text on a single line. Write a program that translates to text to Pig Latin. Pig latin is made by taking the first letter of every word, moving it to the end of the word and adding "ay"

assumptions are: letters are all lowercase, each word is seperated by a single space, numbers remain unchanged, there are no punction marks