Max Difference Challenge
You're given a sequence of numbers on a single line via standard input, each separated by a single space. print the maximum difference (in absolute value) between any two numbers in the sequence on a single line on standard output.

Each number is separated by a single space, and the numbers can be positive, negative or zero.